Former-commit-id: b1f53fa8608ed5662ffde6dc0c18b1ecced5d8e2

+++ dannce/engine/io.py
@@ -49,7 +49,7 @@ def load_com(path):
     d = sio.loadmat(path)["com"]
     data = {}
     data["com3d"] = d["com3d"][0, 0]
-    data["sampleID"] = d["sampleID"][0, 0]
+    data["sampleID"] = d["sampleID"][0, 0].astype(int)
     return data
